Fixing test
